// apps/indexer/src/database/schemas/public/NftCollection.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import { type ColumnType, type Selectable, type Insertable, type Updateable } from 'kysely';

/** Identifier type for public.NftCollection */
export type NftCollectionId = string & { __brand: 'NftCollectionId' };

/** Represents the table public.NftCollection */
export default interface NftCollectionTable {
  id: ColumnType<NftCollectionId, NftCollectionId | undefined, NftCollectionId>;

  address: ColumnType<string, string, string>;

  uri: ColumnType<string, string, string>;

  image: ColumnType<string, string, string>;

  name: ColumnType<string, string, string>;

  description: ColumnType<string, string, string>;

  raw: ColumnType<unknown, unknown, unknown>;
}

export type NftCollection = Selectable<NftCollectionTable>;

export type NewNftCollection = Insertable<NftCollectionTable>;

export type NftCollectionUpdate = Updateable<NftCollectionTable>;