// apps/indexer/src/database/schemas/public/NftAttribute.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import { type NftId } from './Nft';
import { type ColumnType, type Selectable, type Insertable, type Updateable } from 'kysely';

/** Identifier type for public.NftAttribute */
export type NftAttributeId = string & { __brand: 'NftAttributeId' };

/** Represents the table public.NftAttribute */
export default interface NftAttributeTable {
  id: ColumnType<NftAttributeId, NftAttributeId | undefined, NftAttributeId>;

  collectionAddress: ColumnType<string, string, string>;

  nftId: ColumnType<NftId, NftId, NftId>;

  key: ColumnType<string, string, string>;

  value: ColumnType<string, string, string>;

  raw: ColumnType<unknown, unknown, unknown>;
}

export type NftAttribute = Selectable<NftAttributeTable>;

export type NewNftAttribute = Insertable<NftAttributeTable>;

export type NftAttributeUpdate = Updateable<NftAttributeTable>;